Machine Learning/Image Classification/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
scikit-learn